Tests/LibWeb/Text/expected/SVG/svg-foreignobject-abspos-outside-cb-text-change.txt

width-grew=true
position-stable=true
